medicine_website - Copy/requirements.txt
Flask
python-dotenv
requests
groq
beautifulsoup4
lxml